check if item exists before exclude it

+++ inventory_api/src/services/RemoveItemService.ts
@@ -2,6 +2,8 @@ import { inject, injectable } from 'tsyringe';
 
 import ItemRepositoryDTO from '../dto/repositories/ItemRepositoryDTO';
 
+import NewError from '../errors/NewError';
+
 @injectable()
 export default class RemoveItemService {
     constructor(
@@ -10,6 +12,12 @@ export default class RemoveItemService {
     ) {}
 
     public async execute(id: string): Promise<void> {
+        const findItem = await this.itemRepository.findById(id);
+
+        if (!findItem) {
+            throw new NewError('Item not found. Check item id', 404);
+        }
+
         await this.itemRepository.remove(id);
     }
 }
